Working json response with keyfile
