+++ src/constants.rs
@@ -1,0 +1,5 @@
+
+
+// this is the base domain which peach-dyndns creates subdomains under
+// e.g. blue.dyn.peachcloud.org
+pub const BASE_DOMAIN: &str = "dyn.commoninternet.net";
